Move client_logger() back into client module
