+++ src/lib.rs
@@ -88,17 +88,18 @@ mod callback;
 mod userdata;
 mod value;
 
+pub(crate) use self::{
+    client::client_logger,
+    data_type::{data_type, enum_variants},
+    service::{ServiceRequest, ServiceResponse},
+    value::{ArrayValue, NonScalarValue},
+};
 pub use self::{
     client::{Client, ClientBuilder},
     data_type::DataType,
     error::{Error, Result},
     userdata::Userdata,
     value::{ScalarValue, ValueType, VariantValue},
-};
-pub(crate) use self::{
-    data_type::{data_type, enum_variants},
-    service::{ServiceRequest, ServiceResponse},
-    value::{ArrayValue, NonScalarValue},
 };
 
 #[cfg(feature = "tokio")]
